+++ python-stdlib/hashlib-core/manifest.py
@@ -1,0 +1,3 @@
+metadata(version="1.0")
+
+package("hashlib")
