+++ python-stdlib/hashlib-sha224/manifest.py
@@ -1,0 +1,4 @@
+metadata(version="1.0", description="Adds the SHA224 hash algorithm to hashlib.")
+
+require("hashlib-sha256")
+package("hashlib")
