+++ python-stdlib/hashlib-sha256/manifest.py
@@ -1,0 +1,4 @@
+metadata(version="1.0", description="Adds the SHA256 hash algorithm to hashlib.")
+
+require("hashlib-core")
+package("hashlib")
