+++ python-stdlib/hashlib-sha384/manifest.py
@@ -1,0 +1,4 @@
+metadata(version="1.0", description="Adds the SHA384 hash algorithm to hashlib.")
+
+require("hashlib-sha512")
+package("hashlib")
