--- python-stdlib/hashlib/hashlib/_sha224.py
@@ -1,1 +1,0 @@
-from ._sha256 import sha224
