--- python-stdlib/hashlib/hashlib/_sha384.py
@@ -1,1 +1,0 @@
-from ._sha512 import sha384
