+++ python-stdlib/hashlib/manifest.py
@@ -1,3 +1,8 @@
-metadata(version="2.4.0-4")
+metadata(version="2.5.0")
 
-package("hashlib")
+# This is a collection package that gets all hash functions. To save code and
+# memory size, prefer to install just the algorithm you need.
+require("hashlib-sha224")
+require("hashlib-sha256")
+require("hashlib-sha384")
+require("hashlib-sha512")
